Handle energy charge max usage units from imported CSV files
For rate data CSV files with maximum usage units value of -1, reset to zero (kWh) and display UI message box. This may happen when a user used an older version of SAM to import a rate from the URDB with no value for the max usage field and then exported it to a CSV file.

Revise messages in message boxes.