Replace RapidJSON IsDouble() with IsNumber()
In some cases for URDB data, e.g., energy charge tier max value, JSON stores an integer for a value that SAM requires to be double. IsNumber() allows integers to be read in and then be converted to double.

Also changing this for geotools so that integer values of lat/lon are read in from JSON as valid values.